feat: memberInfo에 level,wins 추가로인한 변경사항

+++ backend/src/main/java/com/whatsong/domain/websocket/dto/responseDto/ChannelUserResponseItem.java
@@ -14,7 +14,7 @@ import lombok.ToString;
 @ToString
 @Builder
 public class ChannelUserResponseItem {
-	private int userLevel;
+	private long userLevel;
 	private String nickname;
 	private Boolean isGaming;
 }
